stream-model extracted to own file
