+++ common/CMakeLists.txt
@@ -49,6 +49,8 @@ set(COMMON_SRC
     "${CMAKE_CURRENT_LIST_DIR}/subdevice-model.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/processing-block-model.h"
     "${CMAKE_CURRENT_LIST_DIR}/processing-block-model.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/stream-model.h"
+    "${CMAKE_CURRENT_LIST_DIR}/stream-model.cpp"
     )
 
 set(SW_UPDATE_FILES
