+++ common/viewer.h
@@ -6,6 +6,7 @@
 #include <unordered_set>
 #include "model-views.h"
 #include "subdevice-model.h"
+#include "stream-model.h"
 #include "notifications.h"
 #include "skybox.h"
 #include "measurement.h"
